refactor(preprints-components): Rename and reorganize component files for consistency
