+++ src/app/features/preprints/components/filters/preprints-resources-filters/preprints-resources-filters.component.scss
