+++ src/app/features/preprints/components/filters/preprints-resources/preprints-resources.component.html
@@ -23,7 +23,7 @@
       } @else {
         @if (isAnyFilterOptions()) {
           <img
-            ngSrc="assets/icons/colored/filter.svg"
+            ngSrc="../../../../../../assets/icons/colored/filter.svg"
             tabindex="0"
             role="button"
             alt="filter by"
@@ -34,7 +34,7 @@
           />
         }
         <img
-          ngSrc="assets/icons/colored/sort.svg"
+          ngSrc="../../../../../../assets/icons/colored/sort.svg"
           tabindex="0"
           role="button"
           alt="sort by"
