+++ src/app/features/preprints/components/filters/preprints-resources/preprints-resources.component.scss
