+++ src/app/features/preprints/components/index.ts
@@ -1,15 +1,17 @@
+export { AdvisoryBoardComponent } from './advisory-board/advisory-board.component';
 export { BrowseBySubjectsComponent } from './browse-by-subjects/browse-by-subjects.component';
+export { PreprintsCreatorsFilterComponent } from './filters/preprints-creators-filter/preprints-creators-filter.component';
+export { PreprintsDateCreatedFilterComponent } from './filters/preprints-date-created-filter/preprints-date-created-filter.component';
+export { PreprintsInstitutionFilterComponent } from './filters/preprints-institution-filter/preprints-institution-filter.component';
+export { PreprintsLicenseFilterComponent } from './filters/preprints-license-filter/preprints-license-filter.component';
+export { PreprintProviderFooterComponent } from './preprint-provider-footer/preprint-provider-footer.component';
+export { PreprintProviderHeroComponent } from './preprint-provider-hero/preprint-provider-hero.component';
 export { PreprintServicesComponent } from './preprint-services/preprint-services.component';
-export { AdvisoryBoardComponent } from '@osf/features/preprints/components/advisory-board/advisory-board.component';
-export { PreprintsCreatorsFilterComponent } from '@osf/features/preprints/components/filters/preprints-creators-filter/preprints-creators-filter.component';
-export { PreprintsDateCreatedFilterComponent } from '@osf/features/preprints/components/filters/preprints-date-created-filter/preprints-date-created-filter.component';
-export { PreprintsInstitutionFilterComponent } from '@osf/features/preprints/components/filters/preprints-institution-filter/preprints-institution-filter.component';
-export { PreprintsLicenseFilterComponent } from '@osf/features/preprints/components/filters/preprints-license-filter/preprints-license-filter.component';
-export { PreprintsSubjectFilterComponent } from '@osf/features/preprints/components/filters/preprints-subject/preprints-subject-filter.component';
-export { PreprintsFilterChipsComponent } from '@osf/features/preprints/components/preprints-filter-chips/preprints-filter-chips.component';
-export { PreprintsHelpDialogComponent } from '@osf/features/preprints/components/preprints-help-dialog/preprints-help-dialog.component';
-export { PreprintsResourcesComponent } from '@osf/features/preprints/components/preprints-resources/preprints-resources.component';
-export { PreprintsResourcesFiltersComponent } from '@osf/features/preprints/components/preprints-resources-filters/preprints-resources-filters.component';
-export { FileStepComponent } from '@osf/features/preprints/components/submit-steps/file-step/file-step.component';
-export { MetadataComponent } from '@osf/features/preprints/components/submit-steps/metadata/metadata.component';
-export { TitleAndAbstractStepComponent } from '@osf/features/preprints/components/submit-steps/title-and-abstract-step/title-and-abstract-step.component';
+export { PreprintsHelpDialogComponent } from './preprints-help-dialog/preprints-help-dialog.component';
+export { PreprintsFilterChipsComponent } from '@osf/features/preprints/components/filters/preprints-filter-chips/preprints-filter-chips.component';
+export { PreprintsResourcesComponent } from '@osf/features/preprints/components/filters/preprints-resources/preprints-resources.component';
+export { PreprintsResourcesFiltersComponent } from '@osf/features/preprints/components/filters/preprints-resources-filters/preprints-resources-filters.component';
+export { PreprintsSubjectFilterComponent } from '@osf/features/preprints/components/filters/preprints-subject-filter/preprints-subject-filter.component';
+export { FileStepComponent } from '@osf/features/preprints/components/stepper/file-step/file-step.component';
+export { MetadataStepComponent } from '@osf/features/preprints/components/stepper/metadata-step/metadata-step.component';
+export { TitleAndAbstractStepComponent } from '@osf/features/preprints/components/stepper/title-and-abstract-step/title-and-abstract-step.component';
