+++ src/app/features/preprints/components/preprint-provider-hero/preprint-provider-hero.component.ts
@@ -9,10 +9,11 @@ import { ChangeDetectionStrategy, Component, inject, input, output } from '@angu
 import { FormControl } from '@angular/forms';
 import { RouterLink } from '@angular/router';
 
-import { PreprintsHelpDialogComponent } from '@osf/features/preprints/components';
 import { PreprintProviderDetails } from '@osf/features/preprints/models';
 import { SearchInputComponent } from '@shared/components';
 import { DecodeHtmlPipe } from '@shared/pipes';
+
+import { PreprintsHelpDialogComponent } from '../preprints-help-dialog/preprints-help-dialog.component';
 
 @Component({
   selector: 'osf-preprint-provider-hero',
