+++ src/app/features/preprints/components/stepper/file-step/file-step.component.ts
