+++ src/app/features/preprints/components/stepper/metadata-step/metadata-step.component.html
