+++ src/app/features/preprints/components/stepper/title-and-abstract-step/title-and-abstract-step.component.scss
