+++ src/app/features/preprints/pages/preprint-provider-discover/preprint-provider-discover.component.ts
@@ -17,17 +17,16 @@ import { takeUntilDestroyed, toSignal } from '@angular/core/rxjs-interop';
 import { FormControl } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 
-import { PreprintsResourcesComponent } from '@osf/features/preprints/components';
-import { PreprintProviderHeroComponent } from '@osf/features/preprints/components/preprint-provider-hero/preprint-provider-hero.component';
+import { PreprintProviderHeroComponent, PreprintsResourcesComponent } from '@osf/features/preprints/components';
 import { GetPreprintProviderById, PreprintProvidersSelectors } from '@osf/features/preprints/store/preprint-providers';
 import {
   GetResources,
+  PreprintsDiscoverSelectors,
   ResetState,
   SetProviderIri,
   SetSearchText,
   SetSortBy,
 } from '@osf/features/preprints/store/preprints-discover';
-import { PreprintsDiscoverSelectors } from '@osf/features/preprints/store/preprints-discover/preprints-discover.selectors';
 import {
   PreprintsResourcesFiltersSelectors,
   ResetFiltersState,
