+++ src/app/features/preprints/pages/submit-preprint-stepper/submit-preprint-stepper.component.ts
@@ -19,7 +19,7 @@ import { ActivatedRoute } from '@angular/router';
 
 import {
   FileStepComponent,
-  MetadataComponent,
+  MetadataStepComponent,
   TitleAndAbstractStepComponent,
 } from '@osf/features/preprints/components';
 import { submitPreprintSteps } from '@osf/features/preprints/constants';
@@ -29,13 +29,13 @@ import {
   ResetStateAndDeletePreprint,
   SetSelectedPreprintProviderId,
 } from '@osf/features/preprints/store/submit-preprint';
-import { StepperComponent } from '@shared/components/stepper/stepper.component';
+import { StepperComponent } from '@shared/components';
 import { BrandService } from '@shared/services';
 import { BrowserTabHelper, HeaderStyleHelper, IS_WEB } from '@shared/utils';
 
 @Component({
   selector: 'osf-submit-preprint-stepper',
-  imports: [Skeleton, StepperComponent, TitleAndAbstractStepComponent, FileStepComponent, MetadataComponent],
+  imports: [Skeleton, StepperComponent, TitleAndAbstractStepComponent, FileStepComponent, MetadataStepComponent],
   templateUrl: './submit-preprint-stepper.component.html',
   styleUrl: './submit-preprint-stepper.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
